refactor: move toTitleCase into string standarization file
